// src/permohonan-verifikasi/dto/create.permohonan-verifikasi.dto.ts
export class CreateResponsePermohonanVerifikasiPpnsVerifikasiPpnsDto {
  id: number | null;
  id_data_ppns: number | null;
  id_surat: number | null;
  masa_kerja: any | null;
  pendidikan_terakhir: any | null;
  teknis_operasional_penegak_hukum: boolean | null;
  jabatan: string | null;
  surat_sehat_jasmani_rohani: any | null;
  dp3: any | null;
}

export class CreateResponsePermohonanVerifikasiUploadDokumenPpnsDto {
  message: string;
}